moved diffmap function and label transf function to acdc.tl. Deleted diffmap.py and label_transf.py files
