+++ acdc_py/_pl.py
@@ -6,7 +6,6 @@ from sklearn.metrics import silhouette_samples, silhouette_score
 import seaborn as sns
 import pandas as pd
 import numpy as np
-from ._get_opt import _GS_metric_value
 
 ### ---------- EXPORT LIST ----------
 __all__ = []
@@ -191,6 +190,7 @@ def _metric_vs_n_clusts(
     ss = np.zeros(len(unique_n_clusts))
 
     k = 0
+    from ._get_opt import _GS_metric_value
     for n_clusts in unique_n_clusts:
         ss[k] = _GS_metric_value(adata, n_clusts=n_clusts)
         k+=1
